move Test types to test-result package
