+++ packages/jest-circus/src/legacy-code-todo-rewrite/jestAdapter.ts
@@ -6,9 +6,8 @@
  */
 
 import type {JestEnvironment} from '@jest/environment';
-import type {TestResult} from '@jest/test-result';
+import type {TestFileEvent, TestResult} from '@jest/test-result';
 import type {Config} from '@jest/types';
-import type {TestFileEvent} from 'jest-runner';
 import type Runtime from 'jest-runtime';
 import type {SnapshotStateType} from 'jest-snapshot';
 import {deepCyclicCopy} from 'jest-util';
