+++ packages/jest-circus/tsconfig.json
@@ -10,7 +10,6 @@
     {"path": "../jest-environment"},
     {"path": "../jest-matcher-utils"},
     {"path": "../jest-message-util"},
-    {"path": "../jest-runner"},
     {"path": "../jest-runtime"},
     {"path": "../jest-snapshot"},
     {"path": "../jest-test-result"},
