+++ packages/jest-core/src/ReporterDispatcher.ts
@@ -10,10 +10,10 @@
 import type {Reporter, ReporterOnStartOptions} from '@jest/reporters';
 import type {
   AggregatedResult,
+  Test,
   TestCaseResult,
   TestResult,
 } from '@jest/test-result';
-import type {Test} from 'jest-runner';
 import type {Context} from 'jest-runtime';
 
 export default class ReporterDispatcher {
