+++ packages/jest-core/src/__tests__/SearchSource.test.ts
@@ -7,9 +7,9 @@
  */
 
 import * as path from 'path';
+import type {Test} from '@jest/test-result';
 import type {Config} from '@jest/types';
 import {normalize} from 'jest-config';
-import type {Test} from 'jest-runner';
 import Runtime from 'jest-runtime';
 import SearchSource, {SearchResult} from '../SearchSource';
