+++ packages/jest-runner/src/testWorker.ts
@@ -7,7 +7,11 @@
  */
 
 import exit = require('exit');
-import type {SerializableError, TestResult} from '@jest/test-result';
+import type {
+  SerializableError,
+  TestFileEvent,
+  TestResult,
+} from '@jest/test-result';
 import type {Config} from '@jest/types';
 import HasteMap, {SerializableModuleMap} from 'jest-haste-map';
 import {separateMessageFromStack} from 'jest-message-util';
@@ -15,11 +19,7 @@ import type Resolver from 'jest-resolve';
 import Runtime from 'jest-runtime';
 import {messageParent} from 'jest-worker';
 import runTest from './runTest';
-import type {
-  ErrorWithCode,
-  TestFileEvent,
-  TestRunnerSerializedContext,
-} from './types';
+import type {ErrorWithCode, TestRunnerSerializedContext} from './types';
 
 export type SerializableResolver = {
   config: Config.ProjectConfig;
