+++ packages/jest-runner/src/types.ts
@@ -8,28 +8,15 @@
 import Emittery = require('emittery');
 import type {JestEnvironment} from '@jest/environment';
 import type {
-  AssertionResult,
   SerializableError,
+  Test,
+  TestFileEvent,
   TestResult,
 } from '@jest/test-result';
 import type {Config} from '@jest/types';
-import type {FS as HasteFS, ModuleMap} from 'jest-haste-map';
-import type Resolver from 'jest-resolve';
 import type RuntimeType from 'jest-runtime';
 
 export type ErrorWithCode = Error & {code?: string};
-export type Test = {
-  context: Context;
-  duration?: number;
-  path: Config.Path;
-};
-
-export type Context = {
-  config: Config.ProjectConfig;
-  hasteFS: HasteFS;
-  moduleMap: ModuleMap;
-  resolver: Resolver;
-};
 
 export type OnTestStart = (test: Test) => Promise<void>;
 export type OnTestFailure = (
@@ -40,19 +27,6 @@ export type OnTestSuccess = (
   test: Test,
   testResult: TestResult,
 ) => Promise<void>;
-
-// Typings for `sendMessageToJest` events
-export type TestEvents = {
-  'test-file-start': [Test];
-  'test-file-success': [Test, TestResult];
-  'test-file-failure': [Test, SerializableError];
-  'test-case-result': [Config.Path, AssertionResult];
-};
-
-export type TestFileEvent<T extends keyof TestEvents = keyof TestEvents> = (
-  eventName: T,
-  args: TestEvents[T],
-) => unknown;
 
 export type TestFramework = (
   globalConfig: Config.GlobalConfig,
