+++ packages/jest-test-result/src/index.ts
@@ -24,6 +24,9 @@ export type {
   SnapshotSummary,
   Status,
   Suite,
+  Test,
+  TestEvents,
+  TestFileEvent,
   TestResult,
   TestCaseResult,
   V8CoverageResult,
