+++ packages/jest-test-sequencer/tsconfig.json
@@ -6,7 +6,6 @@
   },
   "references": [
     {"path": "../jest-haste-map"},
-    {"path": "../jest-runner"},
     {"path": "../jest-runtime"},
     {"path": "../jest-test-result"}
   ]
